Renaming MetaDB to SMILE

Refactored packages and classes
