+++ .circleci/config.yml
@@ -10,7 +10,7 @@ orbs:
 jobs:
   run_checkstyle:
     <<: *defaults
-    working_directory: /tmp/repos/cmo-metadb-common
+    working_directory: /tmp/repos/smile-commons
     steps:
       - checkout
       - run:
@@ -21,9 +21,9 @@ jobs:
           command: mvn checkstyle:checkstyle
       - run:
           name: "Validating checkstyle plugin report..."
-          command: bash test/scripts/generate_checkstyle_report.sh /tmp/repos/cmo-metadb-common
+          command: bash test/scripts/generate_checkstyle_report.sh /tmp/repos/smile-commons
       - store_artifacts:
-          path: /tmp/repos/cmo-metadb-common/checkstyle_report_compiled.txt
+          path: /tmp/repos/smile-commons/checkstyle_report_compiled.txt
           destination: /checkstyle_report.txt
 
 workflows:
