+++ pom.xml
@@ -3,8 +3,8 @@
                 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>org.mskcc.cmo</groupId>
-  <artifactId>cmo-metadb-common</artifactId>
-  <name>CMO MetaDB Commons</name>
+  <artifactId>smile-commons</artifactId>
+  <name>CMO SMILE Commons</name>
   <version>1.0.0-SNAPSHOT</version>
   <description>master maven module</description>
   <packaging>jar</packaging>
