+++ src/test/java/org/mskcc/smile/commons/model/MockJsonTestData.java
@@ -1,4 +1,4 @@
-package org.mskcc.cmo.metadb.model;
+package org.mskcc.smile.commons.model;
 
 /**
  *
